A collection of scripts which can be added to any Unity3D game object. These scripts will inteperate whether a mouse has clicked on an object, or a touch from a touchscreen.
The TouchInputController will recieve input from a touch screen. The MouseToTouch will recieve input from the mouse. Both of these scripts will spawn a prefab for each input touching the screen which the TouchMarker script must be a componet of.  The TouchMarker script will store the position of the touch, the position it was created at, and the movement delta of the touch. The gestures script will return a value of the scale of two touches moving in or away from each other which can be used for a zooming affect on the screen.